% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brmsish-package.R
\docType{package}
\name{brmsish}
\alias{brmsish}
\title{brmsish: random stuff on brms bayesian models}
\description{
A poorly documented package with random, tangentially related functions to do visualize or modify brms models.
}
\details{
The main features of the package are:
  \itemize{
  \item Creation of (customized) folder structure
  \item Simplify the inclusion of big data files with version control software and online collaborative platforms (e.g. github)
  }

License: GPL (>= 2)
}
\author{
Marcelo Araya-Salas

Maintainer: Marcelo Araya-Salas (\email{marcelo.araya@ucr.ac.cr})
}
