% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrasts.R
\name{contrasts}
\alias{contrasts}
\title{Estimate hypothesis testing for all pairwise comparisons}
\usage{
contrasts(fit = NULL, predictor, level.sep = " - ", xlab = "Effect size",
gsub.pattern = NULL, gsub.replacement = NULL, n.posterior = 2000,
fill = "#6DCD59FF", sort.levels = NULL, html.table  = FALSE,
read.file = NULL, plot = FALSE, plot.area.prop = 1, highlight = FALSE)
}
\arguments{
\item{fit}{A brmsfit object.}

\item{predictor}{The name of the categorical predictor in the model fit for which contrasts will be computed.}

\item{level.sep}{A character string to separate the levels in the output.}

\item{xlab}{A character string with the horizontal axis label. Default is "Effect size".}

\item{gsub.pattern}{A vector with character strings to be replaced}

\item{gsub.replacement}{A vector with character strings to use for replacement.}

\item{n.posterior}{Number of posterior samples to use for plotting. Default is 2000.}

\item{fill}{Color for posterior distribution fill. Default is "#6DCD59FF".}

\item{sort.levels}{Character vector with the order to be used for levels in the predictor.}

\item{html.table}{Logical to control whether estimate tables are plotted in html format. Useful for creating Rmd or quarto html reports. Is FALSE (default) the table is return as a data frame object.}

\item{read.file}{Character string with the name of the .rds file containing the model fit.}

\item{plot}{Logical to control if posterior distributions of estimates are plotted. Default is FALSE.}

\item{plot.area.prop}{Positive number to control de proportion of the plotting area of posterior distributions that will be included. Default is 1 (the default area included by \code{\link[ggplot2]{ggplot}}). Useful for adding or removing empty space around distributions.}

\item{highlight}{Logical to control if posterior estimates for which the 95\% credible intervals do not overlap with zero are highlighted. Default is FALSE.}
}
\value{
If \code{plot = TRUE} the function returns a ggplot object with the posterior distributions of the comparisons between predictor levels. If \code{html = FALSE} the function will return a data frame with estimates for each comparison, otherwise it will print the estimates in a table in html format.
}
\description{
\code{contrasts} estimates hypothesis testing for all pairwise comparisons of a predictor in brmsfit object.
}
\details{
Estimates hypothesis testing for all pairwise comparisons of levels from a categorical predictor. The function \code{\link[brms]{hypothesis}} is used internally.
}
\examples{
{
# run model
mod <- brm(Petal.Length ~ Species, iris, chains = 1, iter = 500)

# compute constrasts with plot
contrasts(fit = mod, predictor = "Species", html.table = TRUE, plot = TRUE)
# compute constrasts without plot
contrasts(fit = mod, predictor = "Species", html.table = TRUE, plot = FALSE)
}
}
\references{
{
Araya-Salas (2022), brmsish: random stuff on brms bayesian models. R package version 1.0.0.

Paul-Christian Buerkner (2017). brms: An R Package for Bayesian Multilevel Models Using Stan. Journal of Statistical Software, 80(1), 1-28. doi:10.18637/jss.v080.i01
}
}
\seealso{
\code{\link{fit_summary}}, \code{\link{read_summary}}
}
\author{
Marcelo Araya-Salas \email{marcelo.araya@ucr.ac.cr})
}
