% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extended_summary.R
\name{extended_summary}
\alias{extended_summary}
\title{Print a summary of brmsfit results}
\usage{
extended_summary(fit = NULL, gsub.pattern = NULL,
gsub.replacement = NULL,xlab = "Effect size", n.posterior = 2000,
fit.name = NULL, read.file = NULL, plot.area.prop = 1,
remove.intercepts = FALSE, fill = "#6DCD59FF",
trace.palette = viridis::viridis, effects = NULL, save = FALSE,
dest.path = ".", overwrite = FALSE, robust = FALSE,
width = 8, height = "dynamic", highlight = FALSE,
print.name = TRUE)
}
\arguments{
\item{fit}{A brmsfit object.}

\item{gsub.pattern}{A vector with character strings to be replaced}

\item{gsub.replacement}{A vector with character strings to use for replacement.}

\item{xlab}{A character string with the horizontal axis label. Default is "Effect size".}

\item{n.posterior}{Number of posterior samples to use for plotting. Default is 2000.}

\item{fit.name}{Character string to be added as title. If not supplied and 'read.file' is supplied the name is obtained from the file.}

\item{read.file}{Character string with the name of the .rds file containing the model fit.}

\item{plot.area.prop}{Positive number to control de proportion of the plotting area of posterior distributions that will be included. Default is 1 (the default area included by \code{\link[ggplot2]{ggplot}}). Useful for adding or removing empty space around distributions.}

\item{remove.intercepts}{Logical to control if intercepts are included in the output.}

\item{fill}{Color for posterior distribution fill. Default is "#6DCD59FF".}

\item{trace.palette}{Color palette function for trace plotting. Default is \code{\link[viridis]{viridis}}.}

\item{effects}{Character vector with the name of the effects (e.g. predictor variables) to be included. Optional.}

\item{save}{Logical to control if the summary is saved instead of printed. If TRUE a RDS file with the function output (a list) and a jpeg file is saved into 'dest.path'.}

\item{dest.path}{Directory in which to save results (if \code{save = TRUE}). The current working directory is used as default.}

\item{overwrite}{Logical to control if saved results are overwritten. Defaul is FALSE.}

\item{robust}{Logical to control the type of central tendency measure as in \code{\link[brms]{summary.brmsfit}}).}

\item{width}{Width of posterior distribution + trace plots as in \code{\link[ggplot2]{ggsave}}). Default is 8 (in).}

\item{height}{Height of posterior distribution + trace plots as in \code{\link[ggplot2]{ggsave}}). Default is 'dynamic' which means that the height will increase as more panels (predictors) are added.}

\item{highlight}{Logical to control if posterior estimates for which the 95\% credible intervals do not overlap with zero are highlighted. Default is FALSE.}

\item{print.name}{Logical to control if the name of the model fit is printed (when \code{plot = TRUE}).}
}
\value{
If \code{plot = TRUE} the function returns a model fit table, a coefficient table and a posterior distribution halfeye graph. If \code{save = TRUE} this objects are saved as a RDS file along with jpeg file.
}
\description{
\code{extended_summary} prints a summary of brmsfit results.
}
\details{
It prints a summary of brmsfit results. It includes a model fit table, a coefficient table and a posterior distribution halfeye graph next to a chain trace plot. Tables are produce in html format so the output is nicely printed when knit in Rmarkdown files. You might have to add 'results = 'as.is', warning = FALSE and adjust fig.width to chunk options in Rmarkdown documents.
}
\examples{
{
# run model
mod <- brm(Petal.Length ~ Petal.Width + Species, iris, chains = 1,
iter = 500, file = file.path(tempdir(), "mod"))

# print from fit
extended_summary(fit= mod)

# print summary reading the rds file
extended_summary(read.file = file.path(tempdir(), "mod.rds"))
}
}
\references{
{
Araya-Salas (2022), brmsish: miscellaneous functions to customize brms bayesian regression models. R package version 1.0.0.

Paul-Christian Buerkner (2017). brms: An R Package for Bayesian Multilevel Models Using Stan. Journal of Statistical Software, 80(1), 1-28. doi:10.18637/jss.v080.i01
}
}
\seealso{
\code{\link{combine_rds_fits}}
}
\author{
Marcelo Araya-Salas \email{marcelo.araya@ucr.ac.cr})
}
