% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/increase_iterations.R
\name{increase_iterations}
\alias{increase_iterations}
\title{Check brmsfit objects save as RDS files}
\usage{
increase_iterations(
  path = ".",
  fits = list.files(path = path, pattern = ".RDS$", ignore.case = TRUE, full.names =
    TRUE),
  cores = 1,
  pb = TRUE,
  check.rhat = FALSE,
  increase = "2x",
  chains = 4,
  robust = TRUE,
  html = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{path}{Directory in which to look for .rds files.}

\item{fits}{Name of the .rds files to be read. Optional.}

\item{cores}{Number of cores to use for parallelization. Default is 1 (no parallelization).}

\item{pb}{Logical to control if a progress bar is used. Default is TRUE.}

\item{check.rhat}{NOT YET IMPLEMENTED.}

\item{increase}{NOT YET IMPLEMENTED.}

\item{chains}{NOT YET IMPLEMENTED.}

\item{robust}{Logical to control the type of central tendency measure as in \code{\link[brms]{summary.brmsfit}}). NOT YET IMPLEMENTED.}

\item{html}{Logical to control whether results are returned in html format. Useful for creating Rmd or quarto html reports. Is FALSE (default) the table is return as a data frame object. NOT YET IMPLEMENTED.}

\item{verbose}{Logical to control if messages are printed into the console. NOT YET IMPLEMENTED.}
}
\value{
Returns a data frame with a summary of fitted models. Can be used to make sure all models were run with the same parameters (e.g. before combining models). If \code{html = FALSE} the function will return a data frame, otherwise it will print the estimates in a table in html format. The summary includes: prior, formula, number of iterations, number of chain, thinning, warmup, number of parameters, number of divergent transitions, number of rhats higher than 1.05, tail and bulk effective sample sizes and seed.
}
\description{
\code{increase_iterations} checks brmsfit objects save as RDS files.
}
\details{
The function reads all fits saved as rds files in the supplied directory and generates a table listing the parameters used to fit models.
}
\examples{
{
# create directory
dir.create(file.path(tempdir(), "rdss"))
# run 2 models
md1 <- brm(Petal.Length ~ Petal.Width + Species, iris, chains = 1,
iter = 500, file = file.path(tempdir(), "rdss", "md1"))

md2 <- brm(Petal.Length ~ Species, iris, chains = 1,
iter = 500, file = file.path(tempdir(), "rdss", "md2"))

# check fits
increase_iterations(path = file.path(tempdir(), "rdss"))
}
}
\references{
{
Araya-Salas (2022), brmsish: miscellaneous functions to customize brms bayesian regression models. R package version 1.0.0.

Paul-Christian Buerkner (2017). brms: An R Package for Bayesian Multilevel Models Using Stan. Journal of Statistical Software, 80(1), 1-28. doi:10.18637/jss.v080.i01
}
}
\seealso{
\code{\link{extended_summary}}, \code{\link{combine_rds_fits}}
}
\author{
Marcelo Araya-Salas \email{marcelo.araya@ucr.ac.cr})
}
