% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_summary.R
\name{read_summary}
\alias{read_summary}
\title{Read and print the output from 'extended_summary'}
\usage{
read_summary(path = ".", fill = "#6DCD59FF", highlight = FALSE)
}
\arguments{
\item{path}{Directory in which to look for the 'extended_summary' output files.}

\item{fill}{Color for posterior distribution fill. Default is "#6DCD59FF".}

\item{highlight}{Logical to control if posterior estimates for which the 95\% credible intervals do not overlap with zero are highlighted. Default is FALSE.}
}
\value{
If \code{plot = TRUE} the function reads the saved output of \code{\link{extended_summary}} and prints it.
}
\description{
\code{extended_summary} reads and prints the saved output from 'extended_summary'.
}
\details{
It prints a summary of brmsfit results that were previously saved by \code{\link{extended_summary}}. Useful when models are too big so reading them many times is to be avoided. It includes a model fit table, a coefficient table and a posterior distribution halfeye graph next to a chain trace plot. Tables are produce in html format so the output is nicely printed when knit in Rmarkdown files. You might have to add 'results = 'as.is', warning = FALSE and adjust fig.width to chunk options in Rmarkdown documents.
}
\examples{
{
# run model
md1 <- brm(Petal.Length ~ Petal.Width + Species, iris, chains = 1,
iter = 500)

# save summary
extended_summary(fit = md1, save = TRUE, dest.path = tempdir())

# read/print summary
read_summary(path = file.path(tempdir(), "md1"))
}
}
\references{
{
Araya-Salas (2022), brmsish: miscellaneous functions to customize brms bayesian regression models. R package version 1.0.0.

Paul-Christian Buerkner (2017). brms: An R Package for Bayesian Multilevel Models Using Stan. Journal of Statistical Software, 80(1), 1-28. doi:10.18637/jss.v080.i01
}
}
\seealso{
\code{\link{combine_rds_fits}}, \code{\link{extended_summary}}
}
\author{
Marcelo Araya-Salas \email{marcelo.araya@ucr.ac.cr})
}
