% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brmsish-package.R
\docType{package}
\name{brmsish}
\alias{brmsish-package}
\alias{brmsish}
\title{brmsish: miscellaneous functions to customize brms bayesian regression models}
\description{
A poorly documented package with a few random, tangentially related functions to visualize or somehow manipulate brms models.
}
\details{
The main features of the package are:
  \itemize{
  \item Creation of (customized) folder structure
  \item Simplify the inclusion of big data files with version control software and online collaborative platforms (e.g. github)
  }

License: GPL (>= 2)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/maRce10/brmsish}
  \item Report bugs at \url{https://github.com/maRce10/brmsish/issues}
}

}
\author{
Marcelo Araya-Salas

Maintainer: Marcelo Araya-Salas (\email{marcelo.araya@ucr.ac.cr})
}
